+++ sopa/aggregation/aggregation.py
@@ -10,12 +10,7 @@ from spatialdata.models import TableModel
 
 from .._constants import ATTRS_KEY, SopaAttrs, SopaKeys
 from ..io.explorer.utils import str_cell_id
-from ..utils import (
-    add_spatial_element,
-    get_boundaries,
-    get_spatial_element,
-    get_spatial_image,
-)
+from ..utils import add_spatial_element, get_boundaries, get_spatial_element, get_spatial_image
 from . import aggregate_bins, count_transcripts
 from . import aggregate_channels as _aggregate_channels
 
@@ -34,6 +29,7 @@ def aggregate(
     expand_radius_ratio: float | None = None,
     min_transcripts: int = 0,
     min_intensity_ratio: float = 0.1,
+    no_overlap: bool = False,
     key_added: str | None = "table",
 ):
     """Aggregate gene counts and/or channel intensities over a `SpatialData` object to create an `AnnData` table (saved in `sdata["table"]`).
@@ -57,6 +53,7 @@ def aggregate(
         expand_radius_ratio: Ratio to expand the cells polygons for channels averaging. For instance, a ratio of 0.5 expands the shape radius by 50%. If `None` (default), use 1 if we aggregate bins data, and 0 otherwise.
         min_transcripts: Min number of transcripts to keep a cell.
         min_intensity_ratio: Min ratio of the 90th quantile of the mean channel intensity to keep a cell.
+        no_overlap: *Experimental feature*: If `True`, the (expanded) cells will not overlap for channels and bins aggregation.
         key_added: Key to save the table in `sdata.tables`. If `None`, it will be `f"{shapes_key}_table"`.
     """
     assert points_key is None or bins_key is None, "Provide either `points_key` or `bins_key`, not both."
@@ -86,6 +83,7 @@ def aggregate(
         min_transcripts=min_transcripts,
         min_intensity_ratio=min_intensity_ratio,
         gene_column=gene_column,
+        no_overlap=no_overlap,
         key_added=key_added,
     )
 
@@ -154,6 +152,7 @@ class Aggregator:
         min_intensity_ratio: float = 0,
         average_intensities: bool | None = None,  # deprecated argument
         points_key: str | None = None,  # deprecated argument
+        no_overlap: bool = False,
         key_added: str = SopaKeys.TABLE,
     ):
         aggregate_genes, aggregate_channels = self._legacy_arguments(
@@ -175,6 +174,7 @@ class Aggregator:
                     self.shapes_key,
                     self.bins_key,
                     expand_radius_ratio=1 if expand_radius_ratio is None else expand_radius_ratio,
+                    no_overlap=no_overlap,
                 )
             elif not self.already_has_valid_table(key_added):
                 self.table = count_transcripts(
@@ -190,6 +190,7 @@ class Aggregator:
                 image_key=self.image_key,
                 shapes_key=self.shapes_key,
                 expand_radius_ratio=expand_radius_ratio or 0,
+                no_overlap=no_overlap,
             )
 
             if min_intensity_ratio > 0: